+++ src/https/request.ts
@@ -66,7 +66,7 @@ export default class Request {
      * 文章内容
      */
     async content(item:any): Promise<string> {
-        let url = format(this.handler.contentUrl,{list:item.parent.detail,content:item.url});
+        let url = format(this.handler.contentUrl,{list:item.parent.url||"",content:item.url});
         let content = await this.request(url);
         return this.handler.getContent(content);
     }