+++ src/main.ts
@@ -2,6 +2,7 @@
 import * as vscode from "vscode";
 import log from "./utils/log";
 import fs = require('fs');
+import storage from "./storage/storage";
 export interface Callback {
     run: Function
 }
@@ -18,38 +19,37 @@ export class ReadBook {
         log.info('init version:' + this.version);
         this.init();
     }
-
+    
     init() {
-
 
         // 初始化命令
         this.use(import("./commands"));
-        vscode.workspace.workspaceFolders?.map((item,index) => {
+        vscode.workspace.workspaceFolders?.map((item, index) => {
             vscode.workspace.workspaceFile;
             let uri = vscode.Uri.parse(item.uri.path + '/.domain');
             vscode.workspace.findFiles('.domain/*.json').then(res => {
                 if (res.length !== 0) {
 
-                    fs.mkdir(uri.path.replace('/',''),(res)=>{
-                        fs.writeFileSync(uri.path.replace('/','')+'/test.json','{\n'+
-                            '"name": "笔趣阁",'+
-                            '"url": "https://www.xbiquge.so",'+
-                            '"searchUrlChartSet":"gbk",'+
-                            '"searchUrl": "/modules/article/search.php?searchkey={name}",'+
-                            '"parseSearch": {'+
-                              '"list": "#main li",'+
-                              '"url": ".s2 a:href",'+
-                              '"content": ".s2"'+
-                            '},'+
-                            '"parseCatalog": {'+
-                            '  "list": "dd a",'+
-                            '  "url": ":href",'+
-                            '  "content": ""'+
-                            '},'+
-                            '"parseContent": {'+
-                            '  "content": "#content"'+
-                            '}'+
-                          '}');
+                    fs.mkdir(uri.path.replace('/', ''), (res) => {
+                        fs.writeFileSync(uri.path.replace('/', '') + '/test.json', '{\n' +
+                            '"name": "笔趣阁",\n' +
+                            '"url": "https://www.xbiquge.so",\n' +
+                            '"searchUrlChartSet":"gbk",\n' +
+                            '"searchUrl": "/modules/article/search.php?searchkey={name}",\n' +
+                            '"parseSearch": {\n' +
+                            '"list": "#main li",\n' +
+                            '"url": ".s2 a:href",\n' +
+                            '"content": ".s2"\n' +
+                            '},\n' +
+                            '"parseCatalog": {\n' +
+                            '  "list": "dd a",\n' +
+                            '  "url": ":href",\n' +
+                            '  "content": ""\n' +
+                            '},\n' +
+                            '"parseContent": {\n' +
+                            '  "content": "#content"\n' +
+                            '}\n' +
+                            '}');
                     });
                 }
 
@@ -61,6 +61,9 @@ export class ReadBook {
             console.log(e.removed);
             // TODO
         });
+
+        
+
     }
 
     use(fun: any) {