+++ src/providers/book.ts
@@ -49,6 +49,12 @@ export default {
         provider.books = storage.getStorage('books');
         provider.refresh();
         vscode.window.registerTreeDataProvider('books', provider);
+        let select = storage.getStorage('select-book');
+        if (select) {
+            vscode.commands.executeCommand('read-book-status-bar.select-book', {
+                element: select
+            });
+        }
     },
     setItems(contents: any[]) {
         provider.books = contents;